Backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
sqlalchemy>=1.4.23
pydantic>=1.8.2
python-dotenv>=0.19.0
PyPDF2>=2.0.0
pdfminer.six>=20200517
pdf2image>=1.16.0
pytesseract>=0.3.8
httpx>=0.19.0
anthropic>=0.5.0
